requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
